created function to get arbitrageData

+++ src/main/kotlin/com/Lodge/Lodge/Binance/BinanceService.kt
@@ -15,7 +15,7 @@ class BinanceService(
 
         return binanceClient.getPrices().map { binancePriceResponse ->
             priceModel(binancePriceResponse.symbol
-                    , binancePriceResponse.price
+                    , binancePriceResponse.price.toBigDecimal()
                     , binancePriceResponse.exchange)
         }
     }
